+++ tests/Integration/DatabaseTest.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kreait\Firebase\Tests\Integration;
+
+use Kreait\Firebase\Database;
+use Kreait\Firebase\Factory;
+use Kreait\Firebase\ServiceAccount;
+use Kreait\Firebase\Tests\IntegrationTestCase;
+use Kreait\Firebase\Util\JSON;
+
+class DatabaseTest extends DatabaseTestCase
+{
+    public function testWithSanitizableProjectId()
+    {
+        $credentialsPath = self::$fixturesDir.'/test_credentials.json';
+
+        if (!file_exists($credentialsPath)) {
+            self::markTestSkipped();
+        }
+
+        $credentials = JSON::decode(file_get_contents($credentialsPath), true);
+        $credentials['project_id'] = str_replace('-&+§', ':', $credentials['project_id']);
+
+        $serviceAccount = ServiceAccount::fromArray($credentials);
+        $firebase = (new Factory())
+            ->withServiceAccount($serviceAccount)
+            ->create();
+
+        $firebase->getDatabase()->getRules();
+        $this->assertTrue($noExceptionHasBeenThrown = true);
+    }
+}